sharding-jdbc-transaction: move local transaction unit test to sharding-jdbc.
